update more detail comments, refactor test, refactor hash with assembly
