+++ package.json
@@ -51,7 +51,6 @@
   "license": "ISC",
   "dependencies": {
     "openzeppelin-solidity": "^1.12.0",
-    "solc": "0.4.25",
-    "truffle": "^4.1.14"
+    "truffle": "^4.1.15"
   }
 }
